python/requirements.txt
click==6.7
Flask==0.12
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
olefile==0.44
Pillow==4.0.0
redis==2.10.5
Werkzeug==0.11.15
